move useAbilityStep to section, change enemy spell sequences a bit
