+++ game/scripts/vscripts/TutorialGraph/Steps.ts
@@ -208,71 +208,6 @@ export const moveUnit = (unit: tg.StepArgument<CDOTA_BaseNPC>, moveLocation: tg.
 }
 
 /**
- * Creates a tutorial step that orders a unit to use an ability. Completes on ability cast.
- * @param unit The unit to move.
- * @param moveLocation Location to move the unit to.
- * @param completeIfUnitInvalid Optional param that controls whether step should complete if provided unit is invalid or dead. Default value is false
- */
-export const useAbilityStep = (unit: tg.StepArgument<CDOTA_BaseNPC>, target: tg.StepArgument<CDOTA_BaseNPC | Vector>, abilityName: tg.StepArgument<string>, orderType: tg.StepArgument<UnitOrder>) => {
-    let checkTimer: string | undefined = undefined
-
-    const cleanup = () => {
-        if (checkTimer) {
-            Timers.RemoveTimer(checkTimer)
-            checkTimer = undefined
-        }
-    }
-
-    return tg.step((context, complete) => {
-        const actualUnit = tg.getArg(unit, context)
-        const actualTarget = tg.getArg(target, context)
-        const actualAbilityName = tg.getArg(abilityName, context)
-        const actualOrderType = tg.getArg(orderType, context)
-
-        const ability = actualUnit.FindAbilityByName(actualAbilityName)
-
-        let order: ExecuteOrderOptions;
-        if (ability) {
-            if (ability.GetLevel() === 0)
-                ability.SetLevel(1)
-
-            order = {
-                UnitIndex: actualUnit.GetEntityIndex(),
-                OrderType: actualOrderType,
-                AbilityIndex: ability.GetEntityIndex(),
-            }
-        } else {
-            error("Invalid ability name " + abilityName + " for hero " + actualUnit.GetName())
-        }
-
-        if (typeof actualTarget === typeof CDOTA_BaseNPC) {
-            if (orderType === UnitOrder.CAST_TARGET)
-                order.TargetIndex = (actualTarget as CDOTA_BaseNPC).GetEntityIndex()
-            else
-                order.Position = (actualTarget as CDOTA_BaseNPC).GetAbsOrigin()
-        } else {
-            order.Position = (actualTarget as Vector)
-        }
-
-        if (ability.IsCooldownReady())
-            ExecuteOrderFromTable(order)
-        else
-            error("Ability is on cooldown, cannot execute order.")
-
-        const checkAbilityUsed = () => {
-            if (!ability.IsCooldownReady()) {
-                cleanup()
-                complete()
-            } else {
-                checkTimer = Timers.CreateTimer(0.1, () => checkAbilityUsed())
-            }
-        }
-
-        checkAbilityUsed()
-    }, _ => cleanup())
-}
-
-/**
  * Creates a tutorial step that spawns a unit and waits until it dies.
  * @param unitName Name of the unit to spawn.
  * @param spawnLocation Location to spawn the unit at.
